# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BHUVISION \u2014 Agentic Earth Intelligence",
    "description": "An Interactive Vision-Language Assistant for Multimodal Remote Sensing Image Analysis through Text Queries (SIH26167, ISRO). Powered by 9 autonomous specialist agents.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/investigate": {
      "get": {
        "tags": [
          "Investigation"
        ],
        "summary": "List Investigations",
        "description": "List all previous investigations.",
        "operationId": "list_investigations_api_investigate_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvestigationResponse"
                  },
                  "type": "array",
                  "title": "Response List Investigations Api Investigate Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Investigation"
        ],
        "summary": "Start Investigation",
        "description": "Launch a 9-agent investigation on the specified question and satellite imagery.",
        "operationId": "start_investigation_api_investigate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigate/spectral/analyze": {
      "get": {
        "tags": [
          "Investigation"
        ],
        "summary": "Analyze Spectral Index",
        "description": "Computes multispectral remote sensing indices (NDVI / NDWI / NDBI)\nbased on Sentinel-2 optical spectral bands.",
        "operationId": "analyze_spectral_index_api_investigate_spectral_analyze_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Latitude of inspection target",
              "default": 12.9716,
              "title": "Lat"
            },
            "description": "Latitude of inspection target"
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Longitude of inspection target",
              "default": 77.5946,
              "title": "Lon"
            },
            "description": "Longitude of inspection target"
          },
          {
            "name": "index_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Spectral index: ndvi (vegetation), ndwi (water), or ndbi (built-up)",
              "default": "ndvi",
              "title": "Index Type"
            },
            "description": "Spectral index: ndvi (vegetation), ndwi (water), or ndbi (built-up)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Analyze Spectral Index Api Investigate Spectral Analyze Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigate/measure/area": {
      "post": {
        "tags": [
          "Investigation"
        ],
        "summary": "Measure Polygon Area",
        "description": "Computes exact geodesic surface area (Hectares, km\u00b2, Acres) and perimeter (km)\nfor arbitrary polygon coordinates using WGS84 ellipsoidal geometry.",
        "operationId": "measure_polygon_area_api_investigate_measure_area_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolygonMeasurementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Measure Polygon Area Api Investigate Measure Area Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigate/{investigation_id}": {
      "get": {
        "tags": [
          "Investigation"
        ],
        "summary": "Get Investigation",
        "description": "Retrieve the full result, evidence, and audit trace for an investigation.",
        "operationId": "get_investigation_api_investigate__investigation_id__get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigate/{investigation_id}/geojson": {
      "get": {
        "tags": [
          "Investigation"
        ],
        "summary": "Export Investigation Geojson",
        "description": "Exports the investigation spatial polygons, bounding boxes, and findings\nas an RFC 7946 compliant GeoJSON FeatureCollection ready for QGIS, ArcGIS, and ISRO Bhuvan.",
        "operationId": "export_investigation_geojson_api_investigate__investigation_id__geojson_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Investigation Geojson Api Investigate  Investigation Id  Geojson Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeResult": {
        "properties": {
          "has_change": {
            "type": "boolean",
            "title": "Has Change",
            "default": false
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary",
            "default": ""
          },
          "change_regions": {
            "items": {
              "$ref": "#/components/schemas/ChangeRegion"
            },
            "type": "array",
            "title": "Change Regions"
          },
          "change_map_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Map Path"
          },
          "method_used": {
            "type": "string",
            "title": "Method Used",
            "default": ""
          },
          "total_changed_pixels": {
            "type": "integer",
            "title": "Total Changed Pixels",
            "default": 0
          },
          "total_pixels": {
            "type": "integer",
            "title": "Total Pixels",
            "default": 0
          },
          "change_percentage": {
            "type": "number",
            "title": "Change Percentage",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "ChangeResult",
        "description": "Output of the Bi-Temporal Change Agent."
      },
...
      "FusedEvidence": {
        "properties": {
          "primary_answer": {
            "type": "string",
            "title": "Primary Answer"
          },
          "evidence_items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence Items"
          },
          "supporting_summary": {
            "type": "string",
            "title": "Supporting Summary",
            "default": ""
          },
          "contradictions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contradictions"
          },
          "evidence_strength": {
            "$ref": "#/components/schemas/ConfidenceLevel",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "primary_answer"
        ],
        "title": "FusedEvidence",
        "description": "Output of the Evidence Fusion Agent."
      },
...
      "GroundingResult": {
        "properties": {
          "regions": {
            "items": {
              "$ref": "#/components/schemas/ChangeRegion"
            },
            "type": "array",
            "title": "Regions"
          },
          "overlay_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlay Path"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "title": "GroundingResult",
        "description": "Output of the Visual Grounding Agent."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question"
          },
          "imagery_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Imagery Ids"
          },
          "mode": {
            "$ref": "#/components/schemas/InvestigationMode",
            "default": "auto"
          },
          "demo_scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Scenario",
            "description": "If set, use deterministic demo data for this scenario"
          }
        },
        "type": "object",
        "required": [
          "question",
          "imagery_ids"
        ],
        "title": "InvestigationRequest",
        "description": "Request to start an investigation."
      },
      "InvestigationResponse": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "status": {
            "$ref": "#/components/schemas/InvestigationStatus"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "default": ""
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "sensor_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SensorDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "vqa_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VQAResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "change_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChangeResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "grounding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundingResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "fused_evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FusedEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "visual_overlays": {
            "items": {
              "$ref": "#/components/schemas/VisualOverlay"
            },
            "type": "array",
            "title": "Visual Overlays"
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "status",
          "question"
        ],
        "title": "InvestigationResponse",
        "description": "Complete response from an investigation."
      },
      "InvestigationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "planning",
          "validating",
          "routing",
          "analyzing",
          "grounding",
          "fusing",
          "assessing",
          "complete",
          "error"
        ],
        "title": "InvestigationStatus",
        "description": "Current status of an investigation."
      },
...
      "PolygonMeasurementRequest": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "PolygonMeasurementRequest"
      },
      "QueryPlan": {
        "properties": {
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "requires_temporal": {
            "type": "boolean",
            "title": "Requires Temporal",
            "default": false
          },
          "requires_sar": {
            "type": "boolean",
            "title": "Requires Sar",
            "default": false
          },
          "requires_grounding": {
            "type": "boolean",
            "title": "Requires Grounding",
            "default": true
          },
          "suggested_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Tools"
          },
          "investigation_summary": {
            "type": "string",
            "title": "Investigation Summary",
            "default": ""
          },
          "original_question": {
            "type": "string",
            "title": "Original Question",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "task_type"
        ],
        "title": "QueryPlan",
        "description": "Output of the Query Planner Agent."
      },
...
      "SensorDecision": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/SensorDecisionType"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "optical_available": {
            "type": "boolean",
            "title": "Optical Available",
            "default": false
          },
          "sar_available": {
            "type": "boolean",
            "title": "Sar Available",
            "default": false
          },
          "recommendation_confidence": {
            "$ref": "#/components/schemas/ConfidenceLevel",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "SensorDecision",
        "description": "Output of the Sensor Router Agent."
      },
...
      "VQAResult": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "raw_model_output": {
            "type": "string",
            "title": "Raw Model Output",
            "default": ""
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": ""
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "default": ""
          },
          "logits_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Logits Confidence",
            "description": "Model confidence from logits, if available"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "VQAResult",
        "description": "Output of the Remote-Sensing VQA Agent."
      },
...
      "ValidationResult": {
        "properties": {
          "is_valid": {
            "type": "boolean",
            "title": "Is Valid",
            "default": true
          },
          "file_valid": {
            "type": "boolean",
            "title": "File Valid",
            "default": true
          },
          "dimensions_valid": {
            "type": "boolean",
            "title": "Dimensions Valid",
            "default": true
          },
          "crs_valid": {
            "type": "boolean",
            "title": "Crs Valid",
            "default": true
          },
          "metadata_valid": {
            "type": "boolean",
            "title": "Metadata Valid",
            "default": true
          },
          "sensor_compatible": {
            "type": "boolean",
            "title": "Sensor Compatible",
            "default": true
          },
          "temporal_compatible": {
            "type": "boolean",
            "title": "Temporal Compatible",
            "default": true
          },
          "coverage_valid": {
            "type": "boolean",
            "title": "Coverage Valid",
            "default": true
          },
          "alignment_valid": {
            "type": "boolean",
            "title": "Alignment Valid",
            "default": true
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "title": "ValidationResult",
        "description": "Output of the Input & Geo Validation Agent."
      },
...
      "VisualOverlay": {
        "properties": {
          "overlay_type": {
            "type": "string",
            "title": "Overlay Type"
          },
          "coordinates": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates"
          },
          "color": {
            "type": "string",
            "title": "Color",
            "default": "#3B82F6"
          },
          "opacity": {
            "type": "number",
            "title": "Opacity",
            "default": 0.4
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "overlay_type"
        ],
        "title": "VisualOverlay",
        "description": "Visual overlay for the frontend."
      }
    }
  }
}